+++ src/client/main/canvas.ts
@@ -1,31 +1,28 @@
-export const canvas = document.getElementById('game-window') as HTMLCanvasElement || document.createElement('canvas'); // For testing
+export const canvas = document.getElementById('tscratch-main-canvas') as HTMLCanvasElement || document.createElement('canvas'); // For testing
 export const ctx = canvas.getContext('2d')!;
 
 export const penCanvas = document.createElement('canvas');
 export const penCtx = penCanvas.getContext('2d')!;
-penCanvas.id = 'pen-canvas';
-
-export let aspectRatio: number = 16 / 9;
-export let scale: number;
-
-function updateSize() {
-    canvas.width = aspectRatio * scale;
-    canvas.height = scale;
-
-    penCanvas.width = aspectRatio * scale;
-    penCanvas.height = scale;
-}
-
-export function setScale(newScale: number) {
-    scale = newScale;
-    updateSize();
-}
-
-export function setAspectRatio(newAspectRatio: number) {
-    aspectRatio = newAspectRatio;
-    updateSize();
-}
+penCanvas.id = 'tscratch-pen-canvas';
 
 canvas.parentElement?.insertBefore(penCanvas, canvas);
 
-setScale(500);
+const width = 480;
+const height = 360;
+
+function updateSize() {
+    canvas.width = width;
+    canvas.height = height;
+    penCanvas.width = width;
+    penCanvas.height = height;
+}
+
+export function setScale(newScale: number) {
+    const pixelWidth = width * newScale;
+    
+    canvas.style.width = `${pixelWidth}px`;
+    penCanvas.style.width = `${pixelWidth}px`;
+}
+
+updateSize();
+setScale(1.5);